feat: show alerts across all pages
