+++ app/src/main/java/org/javadominicano/controladores/AlertasController.java
@@ -12,6 +12,7 @@ import org.javadominicano.entidades.DatosVelocidad;
 import org.javadominicano.visualizadorweb.entidades.DatosHumedad;
 import org.javadominicano.visualizadorweb.entidades.DatosTemperatura;
 import org.javadominicano.dto.AlertaActivaDTO;
+import org.javadominicano.servicios.AlertasService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -43,6 +44,9 @@ public class AlertasController {
     @Autowired
     private RepositorioDatosPrecipitacion repoPrecipitacion;
 
+    @Autowired
+    private AlertasService alertasService;
+
     @ModelAttribute("umbrales")
     public Umbrales obtenerUmbrales() {
         Umbrales umbrales = new Umbrales();
@@ -63,7 +67,7 @@ public class AlertasController {
         model.addAttribute("alertas", repoAlerta.findAll());
         model.addAttribute("nuevaAlerta", new Alerta());
 
-        List<AlertaActivaDTO> activas = obtenerAlertasActivas();
+        List<AlertaActivaDTO> activas = alertasService.obtenerAlertasActivas();
         model.addAttribute("alertasActivas", activas);
 
         return "alertas";
@@ -87,40 +91,4 @@ public class AlertasController {
         return "redirect:/alertas";
     }
 
-    private List<AlertaActivaDTO> obtenerAlertasActivas() {
-        List<AlertaActivaDTO> lista = new ArrayList<>();
-
-        DatosTemperatura temp = repoTemperatura.findTopByOrderByFechaDesc(PageRequest.of(0, 1)).get(0);
-        DatosHumedad hum = repoHumedad.findTopByOrderByFechaDesc(PageRequest.of(0, 1)).get(0);
-        DatosVelocidad vel = repoVelocidad.findTopByOrderByFechaDesc(PageRequest.of(0, 1)).get(0);
-        DatosPrecipitacion pre = repoPrecipitacion.findTopByOrderByFechaDesc(PageRequest.of(0, 1)).get(0);
-
-        agregarAlertaActiva(lista, repoAlerta.findByNombre("Temperatura"), temp.getTemperatura(), temp.getFecha());
-        agregarAlertaActiva(lista, repoAlerta.findByNombre("Humedad"), hum.getHumedad(), hum.getFecha());
-        agregarAlertaActiva(lista, repoAlerta.findByNombre("VelocidadViento"), vel.getVelocidad(), vel.getFecha());
-        agregarAlertaActiva(lista, repoAlerta.findByNombre("Precipitacion"), pre.getProbabilidad(), pre.getFecha());
-
-        return lista;
-    }
-
-    private void agregarAlertaActiva(List<AlertaActivaDTO> lista, Alerta alerta, double valor, Timestamp fecha) {
-        if (chequearAlerta(alerta, valor)) {
-            AlertaActivaDTO dto = new AlertaActivaDTO();
-            dto.setAlerta(alerta);
-            dto.setValorActual(valor);
-            dto.setFecha(fecha);
-            lista.add(dto);
-        }
-    }
-
-    private boolean chequearAlerta(Alerta alerta, Double valor) {
-        if (alerta == null || valor == null || !alerta.isActiva()) {
-            return false;
-        }
-        if (">".equals(alerta.getOperador())) {
-            return valor > alerta.getUmbral();
-        } else {
-            return valor < alerta.getUmbral();
-        }
-    }
 }
