+++ app/src/main/resources/templates/estaciones.html
@@ -171,6 +171,7 @@
             </div>
             <button type="button" class="btn-add" id="openNew">Nueva Estación</button>
         </div>
+        <div th:replace="fragments/alertasActivas :: alertas"></div>
 
         <div class="table-responsive">
             <table>
