+++ app/src/main/resources/templates/login.html
@@ -144,6 +144,7 @@
     </div>
 
     <h2>Sistema Meteorológico</h2>
+    <div th:replace="fragments/alertasActivas :: alertas"></div>
     <p>Monitoreo de Estaciones en Tiempo Real</p>
     <p class="welcome">Bienvenido de nuevo, por favor inicia sesión.</p>
