+++ app/src/main/resources/templates/reportePreview.html
@@ -48,6 +48,7 @@
     </div>
     <div class="content">
         <h1 th:text="${reporte.titulo}"></h1>
+        <div th:replace="fragments/alertasActivas :: alertas"></div>
         <p th:text="${reporte.estacion}"></p>
         <div class="tablas">
             <div class="tabla">
