+++ app/src/main/resources/templates/reportes.html
@@ -100,6 +100,7 @@
 
     <div class="content">
         <h1>Reportes</h1>
+        <div th:replace="fragments/alertasActivas :: alertas"></div>
         <p class="subtitle">Generación de reportes y análisis</p>
 
         <div class="report-generator">
