+++ app/src/main/resources/templates/tablas.html
@@ -58,6 +58,7 @@
 
   <div class="content">
     <h1>Tablas de la Base de Datos</h1>
+    <div th:replace="fragments/alertasActivas :: alertas"></div>
 
     <div class="tablas">
       <!-- Velocidad -->
